+++ components/client/typescript/src/index.ts
@@ -1,13 +1,92 @@
 import { FastifyInstance } from 'fastify';
-import { EventObserverOptions, ChainhookNodeOptions, buildServer } from './server';
-import { EventObserverPredicateSchema, predicateHealthCheck } from './predicates';
+import { buildServer } from './server';
+import { predicateHealthCheck } from './predicates';
 import { Payload } from './schemas/payload';
-import { Static } from '@fastify/type-provider-typebox';
+import { Static, Type } from '@fastify/type-provider-typebox';
+import { BitcoinIfThisOptionsSchema, BitcoinIfThisSchema } from './schemas/bitcoin/if_this';
+import { StacksIfThisOptionsSchema, StacksIfThisSchema } from './schemas/stacks/if_this';
 
-/** Function type for a Chainhook event callback */
-export type OnPredicatePayloadCallback = (name: string, payload: Payload) => Promise<void>;
+const EventObserverOptionsSchema = Type.Object({
+  /** Event observer host name (usually '0.0.0.0') */
+  hostname: Type.String(),
+  /** Event observer port */
+  port: Type.Integer(),
+  /** Authorization token for all Chainhook payloads */
+  auth_token: Type.String(),
+  /** Base URL that will be used by Chainhook to send all payloads to this event observer */
+  external_base_url: Type.String(),
+  /** Wait for the chainhook node to be available before submitting predicates */
+  wait_for_chainhook_node: Type.Optional(Type.Boolean({ default: true })),
+  /** Validate the JSON schema of received chainhook payloads and report errors when invalid */
+  validate_chainhook_payloads: Type.Optional(Type.Boolean({ default: false })),
+  /** Validate the authorization token sent by the server is correct. */
+  validate_token_authorization: Type.Optional(Type.Boolean({ default: true })),
+  /** Size limit for received chainhook payloads (default 40MB) */
+  body_limit: Type.Optional(Type.Number({ default: 41943040 })),
+  /** Node type: `chainhook` or `ordhook` */
+  node_type: Type.Optional(
+    Type.Union([Type.Literal('chainhook'), Type.Literal('ordhook')], {
+      default: 'chainhook',
+    })
+  ),
+  /**
+   * Directory where registered predicates will be persisted to disk so they can be recalled on
+   * restarts.
+   */
+  predicate_disk_file_path: Type.String(),
+  /**
+   * How often we should check with the Chainhook server to make sure our predicates are active and
+   * up to date. If they become obsolete, we will attempt to re-register them.
+   */
+  predicate_health_check_interval_ms: Type.Optional(Type.Integer({ default: 5000 })),
+});
+/** Chainhook event observer configuration options */
+export type EventObserverOptions = Static<typeof EventObserverOptionsSchema>;
 
-/** Predicate for the Chainhook Event Observer to configure */
+const ChainhookNodeOptionsSchema = Type.Object({
+  /** Base URL where the Chainhook node is located */
+  base_url: Type.String(),
+});
+/** Chainhook node connection options */
+export type ChainhookNodeOptions = Static<typeof ChainhookNodeOptionsSchema>;
+
+/**
+ * Callback that will receive every single payload sent by Chainhook as a result of any predicates
+ * that have been registered.
+ */
+export type OnPredicatePayloadCallback = (payload: Payload) => Promise<void>;
+
+const IfThisThenNothingSchema = Type.Union([
+  Type.Composite([
+    BitcoinIfThisOptionsSchema,
+    Type.Object({
+      if_this: BitcoinIfThisSchema,
+    }),
+  ]),
+  Type.Composite([
+    StacksIfThisOptionsSchema,
+    Type.Object({
+      if_this: StacksIfThisSchema,
+    }),
+  ]),
+]);
+export const EventObserverPredicateSchema = Type.Composite([
+  Type.Object({
+    name: Type.String(),
+    version: Type.Integer(),
+    chain: Type.String(),
+  }),
+  Type.Object({
+    networks: Type.Object({
+      mainnet: Type.Optional(IfThisThenNothingSchema),
+      testnet: Type.Optional(IfThisThenNothingSchema),
+    }),
+  }),
+]);
+/**
+ * Partial predicate definition that allows users to build the core parts of a predicate and let the
+ * event observer fill in the rest.
+ */
 export type EventObserverPredicate = Static<typeof EventObserverPredicateSchema>;
 
 /**
@@ -23,13 +102,13 @@ export type EventObserverPredicate = Static<typeof EventObserverPredicateSchema>
  */
 export class ChainhookEventObserver {
   private fastify?: FastifyInstance;
-  private serverOpts: EventObserverOptions;
-  private chainhookOpts: ChainhookNodeOptions;
+  private observer: EventObserverOptions;
+  private chainhook: ChainhookNodeOptions;
   private healthCheckTimer?: NodeJS.Timer;
 
-  constructor(serverOpts: EventObserverOptions, chainhookOpts: ChainhookNodeOptions) {
-    this.serverOpts = serverOpts;
-    this.chainhookOpts = chainhookOpts;
+  constructor(observer: EventObserverOptions, chainhook: ChainhookNodeOptions) {
+    this.observer = observer;
+    this.chainhook = chainhook;
   }
 
   /**
@@ -43,12 +122,12 @@ export class ChainhookEventObserver {
     callback: OnPredicatePayloadCallback
   ): Promise<void> {
     if (this.fastify) return;
-    this.fastify = await buildServer(this.serverOpts, this.chainhookOpts, predicates, callback);
-    await this.fastify.listen({ host: this.serverOpts.hostname, port: this.serverOpts.port });
-    if (this.serverOpts.predicate_health_check_interval_ms && this.healthCheckTimer === undefined) {
+    this.fastify = await buildServer(this.observer, this.chainhook, predicates, callback);
+    await this.fastify.listen({ host: this.observer.hostname, port: this.observer.port });
+    if (this.observer.predicate_health_check_interval_ms && this.healthCheckTimer === undefined) {
       this.healthCheckTimer = setInterval(() => {
-        void predicateHealthCheck(this.serverOpts, this.chainhookOpts);
-      }, this.serverOpts.predicate_health_check_interval_ms);
+        void predicateHealthCheck(this.observer, this.chainhook);
+      }, this.observer.predicate_health_check_interval_ms);
     }
   }
